+++ src/analytics/collector.ts
@@ -1,6 +1,6 @@
 import { NativeModules } from 'react-native';
 import NativeInstance from '../nativeInstance';
-import { AnalyticsConfig, CustomDataConfig } from './config';
+import { AnalyticsConfig, CustomDataConfig, SourceMetadata } from './config';
 
 const AnalyticsModule = NativeModules.AnalyticsModule;
 
@@ -13,6 +13,11 @@ export class AnalyticsCollector extends NativeInstance<AnalyticsConfig> {
    * Whether the native `AnalyticsCollector` object has been created.
    */
   isInitialized = false;
+
+  /**
+   * The native player id that this analytics collector is attached to.
+   */
+  playerId?: string;
 
   /**
    * Whether the native `AnalyticsCollector` object has been disposed.
@@ -37,6 +42,7 @@ export class AnalyticsCollector extends NativeInstance<AnalyticsConfig> {
     if (!this.isDestroyed) {
       AnalyticsModule.destroy(this.nativeId);
       this.isDestroyed = true;
+      this.playerId = undefined;
     }
   };
 
@@ -47,6 +53,7 @@ export class AnalyticsCollector extends NativeInstance<AnalyticsConfig> {
    * @param playerId - Native Id of the player to attach this collector instance.
    */
   attach = (playerId: string): void => {
+    this.playerId = playerId;
     AnalyticsModule.attach(this.nativeId, playerId);
   };
 
@@ -55,6 +62,7 @@ export class AnalyticsCollector extends NativeInstance<AnalyticsConfig> {
    * nothing happens.
    */
   detach = (): void => {
+    this.playerId = undefined;
     AnalyticsModule.detach(this.nativeId);
   };
 
@@ -74,7 +82,7 @@ export class AnalyticsCollector extends NativeInstance<AnalyticsConfig> {
    * @param customData - Analytics custom data config.
    */
   setCustomData = (customData: CustomDataConfig) => {
-    AnalyticsModule.setCustomData(this.nativeId, customData);
+    AnalyticsModule.setCustomData(this.nativeId, this.playerId, customData);
   };
 
   /**
@@ -83,7 +91,7 @@ export class AnalyticsCollector extends NativeInstance<AnalyticsConfig> {
    * @returns The current custom data config.
    */
   getCustomData = async (): Promise<CustomDataConfig> => {
-    return AnalyticsModule.getCustomData(this.nativeId);
+    return AnalyticsModule.getCustomData(this.nativeId, this.playerId);
   };
 
   /**
@@ -93,5 +101,19 @@ export class AnalyticsCollector extends NativeInstance<AnalyticsConfig> {
    */
   getUserId = async (): Promise<string> => {
     return AnalyticsModule.getUserId(this.nativeId);
+  };
+
+  /**
+   * Adds source metadata for the current source loaded into the player.
+   * This method should be called every time a new source is loaded into the player to ensure
+   * that the analytics data is correct.
+   * @param sourceMetadata - Source metadata to set.
+   */
+  addSourceMetadata = (sourceMetadata: SourceMetadata) => {
+    return AnalyticsModule.addSourceMetadata(
+      this.nativeId,
+      this.playerId,
+      sourceMetadata
+    );
   };
 }